fix(routing): address review findings — warning threshold, non-null assertion, shared test helper
